Move to peerlist/single.go file
